Fix: Pawzone 반응형 대응 및 Intercept Routing 사용 x
